Separate analytics class from resouce.py
